fc_api::backend

Trait Backend

pub trait Backend<Block: BlockT>: Send + Sync {
    // Required methods
    fn block_hash<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ethereum_block_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<Block::Hash>>, String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn transaction_metadata<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ethereum_transaction_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionMetadata<Block>>, String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn log_indexer(&self) -> &dyn LogIndexerBackend<Block>;
    fn latest_block_hash<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Block::Hash, String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn is_indexed(&self) -> bool { ... }
}

The frontier backend interface.

Required Methods

fn block_hash<'life0, 'life1, 'async_trait>( &'life0 self, ethereum_block_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<Block::Hash>>, String>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get the substrate hash with the given ethereum block hash.

fn transaction_metadata<'life0, 'life1, 'async_trait>( &'life0 self, ethereum_transaction_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionMetadata<Block>>, String>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get the transaction metadata with the given ethereum block hash.

fn log_indexer(&self) -> &dyn LogIndexerBackend<Block>

Returns reference to log indexer backend.

fn latest_block_hash<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Block::Hash, String>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the hash of the latest substrate block fully indexed by the backend.

Provided Methods

fn is_indexed(&self) -> bool

Indicate whether the log indexing feature is supported.